+++ plotnine/tests/test_geom_text_label.py
@@ -20,15 +20,15 @@ df = pd.DataFrame({
 
 def test_text_aesthetics():
     p = (ggplot(df, aes(y='y', label='label')) +
-         geom_text(aes('x', label='label'), size=15, hjust='left') +
+         geom_text(aes('x', label='label'), size=15, ha='left') +
          geom_text(aes('x+1', angle='angle'),
-                   size=15, vjust='top', show_legend=False) +
+                   size=15, va='top', show_legend=False) +
          geom_text(aes('x+2', label='label', alpha='z'),
                    size=15, show_legend=False) +
          geom_text(aes('x+3', color='factor(z)'),
                    size=15, show_legend=False) +
          geom_text(aes('x+5', size='z'),
-                   hjust='right', show_legend=False) +
+                   ha='right', show_legend=False) +
          scale_size_continuous(range=(12, 30)) +
          scale_y_continuous(limits=(-0.5, n-0.5)))
 
@@ -38,15 +38,15 @@ def test_text_aesthetics():
 def test_label_aesthetics():
     p = (ggplot(df, aes(y='y', label='label')) +
          geom_label(aes('x', label='label', fill='z'),
-                    size=15, hjust='left', show_legend=False) +
+                    size=15, ha='left', show_legend=False) +
          geom_label(aes('x+1', angle='angle'),
-                    size=15, vjust='top', show_legend=False) +
+                    size=15, va='top', show_legend=False) +
          geom_label(aes('x+2', label='label', alpha='z'),
                     size=15, show_legend=False) +
          geom_label(aes('x+3', color='factor(z)'),
                     size=15, show_legend=False) +
          geom_label(aes('x+5', size='z'),
-                    hjust='right', show_legend=False) +
+                    ha='right', show_legend=False) +
          scale_size_continuous(range=(12, 30)) +
          scale_y_continuous(limits=(-0.5, n-0.5)))