+++ doc/changelog.rst
@@ -3,7 +3,14 @@ Changelog
 
 v0.1.1
 ------
-*(unreleased)*
+
+API Changes
+***********
+
+- :class:`~plotnine.geoms.geom_text` now uses ``ha`` and ``va`` as
+  parameter names for the horizontal and vertical alignment. This
+  is what matplotlib users expect. The previous names ``hjust`` and
+  ``vjust`` are silently accepted.
 
 v0.1.0
 ------